fs: Rename anon_inode_getfile_secure() and anon_inode_getfd_secure()
commit 4f0b9194bc119a9850a99e5e824808e2f468c348 upstream

The call to the inode_init_security_anon() LSM hook is not the sole
reason to use anon_inode_getfile_secure() or anon_inode_getfd_secure().
For example, the functions also allow one to create a file with non-zero
size, without needing a full-blown filesystem.  In this case, you don't
need a "secure" version, just unique inodes; the current name of the
functions is confusing and does not explain well the difference with
the more "standard" anon_inode_getfile() and anon_inode_getfd().

Of course, there is another side of the coin; neither io_uring nor
userfaultfd strictly speaking need distinct inodes, and it is not
that clear anymore that anon_inode_create_get{file,fd}() allow the LSM
to intercept and block the inode's creation.  If one was so inclined,
anon_inode_getfile_secure() and anon_inode_getfd_secure() could be kept,
using the shared inode or a new one depending on CONFIG_SECURITY.
However, this is probably overkill, and potentially a cause of bugs in
different configurations.  Therefore, just add a comment to io_uring
and userfaultfd explaining the choice of the function.

While at it, remove the export for what is now anon_inode_create_getfd().
There is no in-tree module that uses it, and the old name is gone anyway.
If anybody actually needs the symbol, they can ask or they can just use
anon_inode_create_getfile(), which will be exported very soon for use
in KVM.

[Backport Changes]

1. In file “io_uring/io_uring.c” retained the changes from Euler
commit a7d46ae176b8a (io_uring: drop any code related to SCM_RIGHTS),
which dropped support for passing io_uring file descriptors over
SCM_RIGHTS, so in function io_uring_get_file() instead of returning
dropped variable with file pointer, the existing logic was retained where
function directly returns result of anon_inode_create_getfile(). This
approach aligns with the upstream version in merge commit 09d1c6a80f2cf
from v6.8-rc1, which is retained in future kernel releases.

Suggested-by: Christian Brauner <brauner@kernel.org>
Reviewed-by: Christian Brauner <brauner@kernel.org>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jeevan deep J <j.jeevandeep@amd.com>